pyproject.toml
[tool.poetry]
name = "goz-demo"
version = "0.1.0"
description = ""
authors = ["Maksat Hemdemow <thehighestintheroom00@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.4"
django-redis = "^5.3.0"
python-decouple = "^3.8"
cython = "^3.0.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
pydub = "^0.25.1"
torch = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
